#FRONTEND-167 - Reduced the login-handler time from 2s+io to 1s+io. Refactored out unused Response model.
