+++ app/src/main/java/io/hobaskos/event/eventapp/data/model/enumeration/SocialType.java
@@ -1,4 +1,4 @@
-package io.hobaskos.event.eventapp.data.model;
+package io.hobaskos.event.eventapp.data.model.enumeration;
 
 /**
  * Created by hansp on 01.03.2017.
